Feat: Display today's schedule for lecturers on home screen
This commit introduces the functionality to display the lecturer's (dosen) schedule for the current day on the home screen. It involves adding new methods to the repository and use case layers to fetch this data, updating the database seeding to include a relevant class for today, and implementing the UI logic in `HomeFragment` and `HomeViewModel` for the `dosen` module.

**Key changes:**

-   **Core Module (`core`):**
    -   **`IVirtualClassRepository.kt` & `VirtualClassInteractor.kt`:**
        -   Added `getTodayScheduleDosen(nidn: String): Flow<Resource<List<Kelas>>>` function to fetch today's schedule specifically for a lecturer.
    -   **`VirtualClassRepository.kt`:**
        -   Implemented `getTodayScheduleDosen` to filter all classes by the lecturer's NIDN and the current day of the week.
    -   **`VirtualClassDatabase.kt` (Seed Data):**
        -   Added a new class "Etika Profesi TI" scheduled for the current day under `dosen1Nidn` to ensure testable data.
        -   Modified an existing class name from "Workshop Pemrograman Python" to "Pemrograman Python" and its description.
        -   Removed commented-out section headers like `// Kelas`, `// Enrollments`, etc.

-   **Dosen Module (`dosen`):**
    -   **`HomeViewModel.kt`:**
        -   Injected `VirtualClassUseCase` instead of `IVirtualClassRepository`.
        -   Added `_todayScheduleState` (MutableStateFlow) and `todayScheduleState` (StateFlow) to hold and expose the today's schedule data.
        -   Implemented `fetchTodaySchedule()` to retrieve the schedule using `virtualClassUseCase.getTodayScheduleDosen()` based on the logged-in lecturer's NIDN from `appPreference`.
        -   Called `fetchTodaySchedule()` in the `init` block.
    -   **`HomeFragment.kt`:**
        -   Added `observeTodaySchedule()` to observe `homeViewModel.todayScheduleState`.
        -   Updated UI elements (`progressBarSchedule`, `ivDoesntHaveAnSchedule`, `tvDoesntHaveAnSchedule`, `tvScheduleClassroom`, `tvTimeScheduleHome`, `tvSubjectScheduleHome`, `btnScheduleDetailHome`) based on the loading, success (with data or empty), or error state of the schedule data.
        -   If a schedule exists for today, the first class's room, time, and subject are displayed.
        -   Called `observeTodaySchedule()` in `onViewCreated`.

+++ authentication/src/main/java/com/mjs/authentication/presentation/login/dosen/LoginDosenViewModel.kt
@@ -36,7 +36,7 @@ class LoginDosenViewModel(
                 _loginResult.value = LoginResult.Error("NIDN and password cannot be empty")
                 return@launch
             }
-            val dosen = authDao.loginDosen(nidn, password)
+            val dosen = authDao.loginDosen(nidn.toInt(), password)
             if (dosen != null) {
                 appPreference.saveLoginSession(dosen.nidn, AppPreference.USER_TYPE_DOSEN)
                 _loginResult.value = LoginResult.Success(AppPreference.USER_TYPE_DOSEN)
